Moved logic of `useSendQueries` to `utils`
